[vm] split out instructions and vm frame process from PikaVM.c
