+++ src/vm/inst/inst_glb.c
@@ -1,0 +1,47 @@
+/*
+ * MIT License
+ *
+ * Copyright (c) 2024 azhao 余钊炜 yuzhaowei2002@outlook.com
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a
+ * copy of this software and associated documentation files (the "Software"),
+ * to deal in the Software without restriction, including without limitation
+ * the rights to use, copy, modify, merge, publish, distribute, sublicense,
+ * and/or sell copies of the Software, and to permit persons to whom the
+ * Software is furnished to do so, subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included in
+ * all copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL
+ * THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+ * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING
+ * FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER
+ * DEALINGS IN THE SOFTWARE.
+ */
+#include "inst_common.h"
+
+Arg *vm_inst_handler_GLB(PikaObj *self, PikaVMFrame *vm, char *data, Arg *arg_ret_reg)
+{
+    Arg *global_list_buff = NULL;
+    char* global_list = args_getStr(vm->locals->list, "@g");
+    /* create new global_list */
+    if (NULL == global_list) {
+        args_setStr(vm->locals->list, "@g", data);
+        obj_setFlag(vm->locals, OBJ_FLAG_GLOBALS);
+        goto __exit;
+    }
+    /* append to exist global_list */
+    global_list_buff = arg_newStr(global_list);
+    global_list_buff = arg_strAppend(global_list_buff, ",");
+    global_list_buff = arg_strAppend(global_list_buff, data);
+    args_setStr(vm->locals->list, "@g", arg_getStr(global_list_buff));
+    goto __exit;
+__exit:
+    if (NULL != global_list_buff) {
+        arg_deinit(global_list_buff);
+    }
+    return NULL;
+}
